fetches alerts when page opens
